Added view product

+++ source/backend/dal/Repositories/Interfaces/IProductRepository.cs
@@ -8,6 +8,6 @@ namespace Pims.Dal.Repositories
     /// </summary>
     public interface IProductRepository : IRepository<PimsProduct>
     {
-        IList<PimsProduct> GetByProject(int projectId);
+        IList<PimsProduct> GetByProject(long projectId);
     }
 }
