+++ source/backend/dal/Repositories/ProductRepository.cs
@@ -28,7 +28,7 @@ namespace Pims.Dal.Repositories
         /// </summary>
         /// <param name="projectId"></param>
         /// <returns></returns>
-        public IList<PimsProduct> GetByProject(int projectId)
+        public IList<PimsProduct> GetByProject(long projectId)
         {
             return this.Context.PimsProducts.AsNoTracking()
                 .Where(p => p.ParentProjectId == projectId)
